feat(playground): refactoring file uploader and voice recorder.
