+++ packages/playground-ui/src/structures/IAutoBePlaygroundFileContent.ts
@@ -1,0 +1,13 @@
+import {
+  AutoBeUserMessageAudioContent,
+  AutoBeUserMessageFileContent,
+  AutoBeUserMessageImageContent,
+} from "@autobe/interface";
+
+export interface IAutoBePlaygroundFileContent {
+  file: File;
+  content:
+    | AutoBeUserMessageAudioContent
+    | AutoBeUserMessageFileContent
+    | AutoBeUserMessageImageContent;
+}
